refactored application structure
