+++ src/components/app.js
@@ -4,30 +4,23 @@ import {
     Switch,
     Route
 } from 'react-router-dom';
-import Header from './header';
-import Navigation from './navigation';
-import ProductsContainer from './products-container';
-import Product from "./product";
-import '../styles/app.scss';
-
+import Products from './products';
+import Product from './product';
+import '../styles/index.scss';
 
 function App() {
     return (
-        <div className="app container-fluid">
-            <Header/>
-            <Navigation/>
-            <div className="body-container">
-                <Router>
-                    <Switch>
-                        <Route exact path="/">
-                            <ProductsContainer />
-                        </Route>
-                        <Route path="/product/:id">
-                            <Product />
-                        </Route>
-                    </Switch>
-                </Router>
-            </div>
+        <div className="app">
+            <Router>
+                <Switch>
+                    <Route exact path="/">
+                        <Products />
+                    </Route>
+                    <Route path="/product/:id">
+                        <Product />
+                    </Route>
+                </Switch>
+            </Router>
         </div>
     );
 }
