--- src/styles/app.scss
@@ -1,10 +1,0 @@
-.app {
-    margin: 0;
-    padding: 0;
-}
-
-.body-container {
-    padding: 15px;
-    display: flex;
-    flex-direction: row;
-}
