--- src/styles/filter.scss
@@ -1,21 +1,0 @@
-.filter-list-container {
-    font-family: Arial, Helvetica, sans-serif;
-
-    .filter-wrap {
-        margin: 7px 15px;
-
-        .filter-wrap-title {
-            margin-bottom: 5px;
-            font-weight: bold;
-        }
-
-        .filter-list {
-            list-style: none;
-            padding: 0;
-
-            .filter-element {
-                font-size: 12px;
-            }
-        }
-    }
-}
