--- src/styles/header.scss
@@ -1,22 +1,0 @@
-.header {
-    display: flex;
-    justify-content: space-between;
-    position: sticky;
-    top: 0;
-    height: 80px;
-    padding: 20px 20px 20px 25px;
-    background-color: #e60012;
-
-    .search-container {
-        margin-right: 50px;
-        padding-right: 15px;
-        border-right: 1px solid white;
-
-        .search {
-            padding-top: 4px;
-            width: 22px;
-            height: 22px;
-            color: white;
-        }
-    }
-}
