+++ src/styles/index.scss
@@ -1,0 +1,5 @@
+@import "~bootstrap/dist/css/bootstrap.min.css";
+
+html, body, #root, .app {
+  height: 100%;
+}
