--- src/styles/navigation.scss
@@ -1,57 +1,0 @@
-.navigation {
-    position: sticky;
-    top: 80px;
-    width: 75px;
-    background-color: #e60012;
-    float: left;
-
-    .navigation-container {
-        display: flex;
-        flex-direction: column;
-        padding: 0;
-
-        .navigation-element {
-            width: 72px;
-            color: white;
-            padding: 3px 8px;
-            text-align: center;
-            cursor: pointer;
-            font-size: 12px;
-            font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
-
-            .icon {
-                font-size: 40px;
-            }
-        }
-
-        .navigation-element:hover {
-            background-color: #e60012;
-        }
-
-        .navigation-social {
-            height: 48px;
-            padding: 8px;
-            background-color: #e60012;
-            margin-top: 1px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            cursor: pointer;
-
-            .nav-social-icon {
-                color: white;
-                font-size: 30px;
-            }
-        }
-
-        .navigation-social:hover {
-            background-color: rgb(212, 0, 18);
-        }
-    }
-
-    .navigation-footer {
-        background-color: rgb(230, 0, 18);
-        height: 25px;
-        margin-top: 1px;
-    }
-}
