--- src/styles/pagination-element.scss
@@ -1,9 +1,0 @@
-.button {
-    background-color: #4CAF50;
-    border: none;
-    text-align: center;
-    text-decoration: none;
-    font-size: 16px;
-    margin: 4px 2px;
-    cursor: pointer;
-}
