--- src/styles/product.scss
@@ -1,32 +1,0 @@
-.product-container {
-    position: absolute;
-    top: 72px;
-    left: 88px;
-    overflow-x: hidden;
-    display: flex;
-    flex-direction: row;
-    background-color: white;
-    margin: 10px 0 0 10px;
-    width: 1000px;
-    justify-content: space-between;
-
-    .photo-wrap {
-        width: 500px;
-        height: 500px;
-        .img {
-            height: 100%;
-            width: 100%;
-            background-position: center;
-            background-repeat: no-repeat;
-            background-size: cover;
-        }
-    }
-    .product-content {
-        display: flex;
-        flex-direction: column;
-        justify-content: space-evenly;
-        margin-right: 50px;
-        align-items: center;
-        max-width: 300px;
-    }
-}
