--- src/styles/products-container.scss
@@ -1,3 +1,0 @@
-.products-container {
-    display: flex;
-}
